小调式，注释一些函数

+++ src/main/java/cn/net/wangchenyu/finance/dao/RepairListDao.java
@@ -1,8 +1,10 @@
 package cn.net.wangchenyu.finance.dao;
+
 import cn.net.wangchenyu.finance.model.RepairList;
 import org.springframework.data.repository.CrudRepository;
+
 /**
  * Created by cheneyveron on 7/7/16.
  */
-public interface RepairListDao extends CrudRepository<RepairList,Long> {
+public interface RepairListDao extends CrudRepository<RepairList,Integer> {
 }
